Fixed Issue where invalid file names could be set on MacOS.
There is a bug with Mono's version of Path.GetInvalidFileNameChars() which on windows it returns the right chars but on Mac it only returns a forward slash. This might be find for the OS but not for Unity. I have made a new implementation that is constant regardless of platform.