Added new helper functions to test paths
The goal is to not have the user have to pick which type of path they need. When you are in a Unity directory we want to use different optoins then if we are outside. In the current state the users has to pick the type or get errors. This is going to be done automaticlly soon.